createAbsMv集成内存网络完成;

+++ SMG_NothingIsAll/AIFoundation/AINet/AINet.h
@@ -56,7 +56,13 @@
 -(AIPort*) getNetNodePointersFromDirectionReference_Single:(NSString*)mvAlgsType direction:(MVDirection)direction;
 -(NSArray*) getNetNodePointersFromDirectionReference:(NSString*)mvAlgsType direction:(MVDirection)direction limit:(int)limit;
 -(NSArray*) getNetNodePointersFromDirectionReference:(NSString*)mvAlgsType direction:(MVDirection)direction filter:(NSArray*(^)(NSArray*))filter;
--(void) setNetNodePointerToDirectionReference:(AIKVPointer*)cmvNode_p mvAlgsType:(NSString*)mvAlgsType direction:(MVDirection)direction difStrong:(int)difStrong;
+
+/**
+ *  MARK:--------------------cmvNode或absCMVNode构建时,报告directionReference--------------------
+ *  @param difStrong : mv的迫切度越高,越强;
+ *  @param direction : 方向(delta的正负)
+ */
+-(void) setNetNodePointerToDirectionReference:(AIKVPointer*)cmvNode_p mvAlgsType:(NSString*)mvAlgsType direction:(MVDirection)direction difStrong:(NSInteger)difStrong;
 
 
 //MARK:===============================================================
